+++ src/error/user-error.ts
@@ -1,0 +1,3 @@
+import { NestedError } from "./nested-error";
+
+export class UserError extends NestedError {}